6.2c - update libs with config : move optimConfig.h et consequences
