+++ src/modeler/CMakeLists.txt
@@ -1,7 +1,6 @@
 add_subdirectory(loadFiles)
 add_subdirectory(modeler)
 add_subdirectory(parameters)
-add_subdirectory(optimConfig)
 
 OMESSAGE("  :: modeler")
