+++ src/modeler/loadFiles/readOptimConfig.cpp
@@ -23,7 +23,6 @@
 
 #include <antares/io/file.h>
 #include <antares/logs/logs.h>
-#include <antares/modeler/optimConfig/optimConfig.h>
 #include "antares/io/inputs/yml-optim-config/converter.h"
 #include "antares/io/inputs/yml-optim-config/parser.h"
 #include "antares/solver/modeler/loadFiles/loadFiles.h"
