--- src/modeler/optimConfig/CMakeLists.txt
@@ -1,13 +1,0 @@
-add_library(modeler-optim-config
-            include/antares/modeler/optimConfig/optimConfig.h)
-
-add_library(Antares::modeler-optim-config ALIAS modeler-optim-config)
-
-target_link_libraries(modeler-optim-config
-        PRIVATE
-        Antares::exception
-)
-
-target_include_directories(modeler-optim-config
-        PUBLIC
-	$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>)
