+++ src/solver/utils/ortools_utils.cpp
@@ -299,7 +299,7 @@ const std::map<std::string, struct OrtoolsUtils::SolverNames> OrtoolsUtils::mpSo
 const std::map<std::string, math_opt::SolverType> OrtoolsUtils::mathoptSolverMap = {
   {"pdlp", math_opt::SolverType::kPdlp},
   {"scip", math_opt::SolverType::kGscip},
-  {"xpress", math_opt::SolverType::kXpress},
+  // {"xpress", math_opt::SolverType::kXpress},
   {"gurobi", math_opt::SolverType::kGurobi}};
 
 std::list<std::string> availableLinearSolversList()
