+++ src/study/system-model/CMakeLists.txt
@@ -36,7 +36,6 @@ target_include_directories(antares-study-system-model
 target_link_libraries(antares-study-system-model
         PUBLIC
         Antares::expressions
-        Antares::modeler-optim-config
         PRIVATE
         Antares::exception
 )
