+++ src/study/system-model/include/antares/study/system-model/variable.h
@@ -23,8 +23,8 @@
 #include <string>
 
 #include <antares/expressions/expression.h>
-#include "antares/modeler/optimConfig/optimConfig.h"
 
+#include "optimConfig.h"
 #include "timeAndScenarioType.h"
 #include "valueType.h"
