+++ src/tests/src/solver/optim-model-filler/component-filler-utils/objectives-creators.h
@@ -3,7 +3,6 @@
 #include "antares/expressions/expression.h"
 #include "antares/expressions/nodes/ExpressionsNodes.h"
 #include "antares/modeler-optimisation-container/TimeIndex.h"
-#include "antares/modeler/optimConfig/optimConfig.h"
 #include "antares/study/system-model/objective.h"
 
 using namespace Antares;
